move some image funcs / constants to shared locations
