+++ luxtest_utils.py
@@ -39,3 +39,25 @@ def get_renders_root() -> str:
 
 def get_render_dirs() -> Tuple[str, ...]:
     return tuple(x.name for x in os.scandir(get_renders_root()) if x.is_dir())
+
+
+def get_image_path(light_name, renderer: str, frame: int, ext: str, prefix="", renders_root=""):
+    if not renders_root:
+        renders_root = get_renders_root()
+    ext = ext.lstrip(".")
+    filename = f"{prefix}{light_name}-{renderer}.{frame:04}.{ext}"
+    if ext == "png":
+        base_dir = luxtest_const.WEB_IMG_ROOT
+    elif ext == "exr":
+        base_dir = os.path.join(renders_root, renderer)
+    else:
+        raise ValueError(f"unrecognized extension: {ext}")
+    return os.path.join(base_dir, filename)
+
+
+def get_image_url(light_name, renderer: str, frame: int, ext: str, prefix="", renders_root=""):
+    if not renders_root:
+        renders_root = get_renders_root()
+    image_path = get_image_path(light_name, renderer, frame, ext, prefix=prefix, renders_root=renders_root)
+    rel_path = os.path.relpath(image_path, luxtest_const.WEB_ROOT)
+    return rel_path.replace(os.sep, "/")
